Relocated And Polished Code
https://remorse-studio.atlassian.net/browse/AE-13
