+++ Engine/Source/EntryPoint.cpp
@@ -5,7 +5,7 @@
 //	EXPECT_EQ(0, 0);
 //}
 #define CATCH_CONFIG_MAIN
-#include "../../Externals/catch.hpp"
+#include "catch.hpp"
 
 #define CATCH_CONFIG_MAIN
 #define UNIT_TESTING 1
