+++ Engine/Source/actapch.cpp
@@ -1,10 +1,14 @@
 #include "actapch.h"
-#include "../catch.hpp"
+#include "catch.hpp"
 
-TEST_CASE("FileManager", "FileManagerTesting")
+TEST_CASE("SampleTest", "SampleTestOnly")
 {
-	SECTION("FileCountTest")
+	SECTION("FirstCase")
 	{
 		REQUIRE(0 == 0);
+	}
+	SECTION("SecondCase")
+	{
+		REQUIRE(false == false);
 	}
 }
